tg_bot/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1
structlog==24.1.0
redis==5.0.1